Error Impurity Function improvements and cleanup
git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C541

+++ RoughSets/Infovision.Datamining.RoughSet/InformationMeasure.cs
@@ -259,7 +259,7 @@ namespace Infovision.Datamining.Roughset
                     if (majorCount > maxDecisionProbability)
                         maxDecisionProbability = majorCount;
                 }
-                result += (double)maxDecisionProbability;
+                result += maxDecisionProbability;
             }
             return result / (double)reduct.DataStore.NumberOfRecords;
         }
@@ -362,7 +362,7 @@ namespace Infovision.Datamining.Roughset
         }
 
         public override string Description()
-        {
+        {   
             return "ObjectWeights";
         }
